# Improve the latency benchmark