--- a/benchmarks/fast_sin_cos_2π_benchmark.cpp
+++ b/benchmarks/fast_sin_cos_2π_benchmark.cpp
@@ -13,38 +13,41 @@ namespace principia {
 namespace numerics {
 namespace {
 
-static const __m128d mantissa_bits =
-    _mm_castsi128_pd(_mm_cvtsi64_si128(0x000F'FFFF'FFFF'FFFF));
-// When iterated, the quadrant of the result is unbiased; the latency is
-// between 1 and 2 cycles: at worst this is an and and an xor, at best the xor
-// can only be computed given both trigonometric lines.
-double MixTrigonometricLines(double cos_2πx, double sin_2πx) {
+// Returns a value which has a data dependency on both cos_2πx and sin_2πx.
+// The result is (cos_2πx bitand cos_mask) bitxor sin_2πx.
+// The latency is between 1 and 2 cycles: at worst this is an and and an xor, at
+// best the xor can only be computed given both trigonometric lines.
+double MixTrigonometricLines(double cos_2πx, double sin_2πx, __m128d const cos_mask) {
   __m128d const cos_mantissa_bits =
-      _mm_and_pd(_mm_set_sd(cos_2πx), mantissa_bits);
+      _mm_and_pd(_mm_set_sd(cos_2πx), cos_mask);
   __m128d const sin_all_bits = _mm_set_sd(sin_2πx);
   __m128d const mixed_bits = _mm_xor_pd(cos_mantissa_bits, sin_all_bits);
   return _mm_cvtsd_f64(mixed_bits);
 }
-static const __m128d mantissa_bits_and_4_bits_of_exponent =
-    _mm_castsi128_pd(_mm_cvtsi64_si128(0x00FF'FFFF'FFFF'FFFF));
-// Same as above, but when iterated, the result is quickly confined to the
-// principal quadrant.
+
+static const __m128d mantissa_bits =
+    _mm_castsi128_pd(_mm_cvtsi64_si128(0x000F'FFFF'FFFF'FFFF));
+// When iterated, the quadrant of the result is unbiased.
+double ThoroughlyMixTrigonometricLines(double cos_2πx, double sin_2πx) {
+  return MixTrigonometricLines(cos_2πx, sin_2πx, mantissa_bits);
+}
+
+static const __m128d mantissa_bits_and_5_bits_of_exponent =
+    _mm_castsi128_pd(_mm_cvtsi64_si128(0x01FF'FFFF'FFFF'FFFF));
+// Same as above, but when iterated, the result is quickly confined to [0, 1/8].
 double PoorlyMixTrigonometricLines(double cos_2πx, double sin_2πx) {
-  __m128d const cos_mantissa_bits =
-      _mm_and_pd(_mm_set_sd(cos_2πx), mantissa_bits_and_4_bits_of_exponent);
-  __m128d const sin_all_bits = _mm_set_sd(sin_2πx);
-  __m128d const mixed_bits = _mm_xor_pd(cos_mantissa_bits, sin_all_bits);
-  return _mm_cvtsd_f64(mixed_bits);
+  return MixTrigonometricLines(
+      cos_2πx, sin_2πx, mantissa_bits_and_5_bits_of_exponent);
 }
 
 }  // namespace
 
-void BM_FastSinCos2πLatency(benchmark::State& state) {
+void BM_FastSinCos2πPoorlyPredictedLatency(benchmark::State& state) {
   while (state.KeepRunning()) {
     double sin = π;
     double cos = 0.0;
     for (int i = 0; i < 1e3; ++i) {
-      FastSinCos2π(MixTrigonometricLines(cos, sin), sin, cos);
+      FastSinCos2π(ThoroughlyMixTrigonometricLines(cos, sin), sin, cos);
     }
   }
 }
@@ -78,7 +81,7 @@ void BM_FastSinCos2πThroughput(benchmark::State& state) {
   }
 }
 
-BENCHMARK(BM_FastSinCos2πLatency);
+BENCHMARK(BM_FastSinCos2πPoorlyPredictedLatency);
 BENCHMARK(BM_FastSinCos2πWellPredictedLatency);
 BENCHMARK(BM_FastSinCos2πThroughput);